🎨 Custom hooksを使用してうまくまとめました
