+++ src/hooks/useCounter.jsx
@@ -1,0 +1,18 @@
+import { useCallback, useState } from "react";
+
+export const useCounter = () => {
+    const [count, setCount] = useState(1);
+    const [isShow, setIsShow] = useState(true);
+
+    const handleClick = useCallback(() => {
+        if (count < 10) {
+        setCount((prevCount) => prevCount + 1);
+        }
+    }, [count]);
+
+    const handleDisplay = useCallback(() => {
+        setIsShow((prevIsShow) => !prevIsShow);
+    }, []);
+
+    return { count, isShow, handleClick, handleDisplay };
+};
