+++ src/pages/index.jsx
@@ -1,65 +1,26 @@
 import { Inter } from 'next/font/google'
 import Main from '@/src/components/main/Main'
-import { useCallback, useEffect, useState } from 'react';
+import { useCounter } from '@/src/hooks/useCounter';
+import { useInputArray } from '@/src/hooks/useInputArray';
+import { useBgLightBlue } from '@/src/hooks/useBgLightBlue';
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ['latin'] });
 
 export default function Home() {
-  const [count, setCount] = useState(1);
-  const [text, setText] = useState("");
-  const [inputAlert, setInputAlert] = useState("");
-  const [isShow, setIsShow] = useState(true);
-  const [array, setArray] = useState([]);
-
-  const handleClick = useCallback(() => {
-    if (count < 10) {
-      setCount((prevCount) => prevCount + 1);
-    }
-  }, [count]);
-  
-  const handleDisplay = useCallback(() => {
-    setIsShow((prevIsShow) => !prevIsShow);
-  }, []);
-
-  const handleChange = useCallback((e) => {
-    if(e.target.value.length > 5){
-      setInputAlert("5文字以内にしてね");
-      return;
-    };
-    setInputAlert("");
-    setText(e.target.value.trim());
-  }, []);
-
-  const handleAdd = useCallback(() => {
-    setArray((prevArray) => {
-      if (prevArray.some(item => item === text)) {
-        alert("exiting same items");
-        return prevArray;
-      }
-      return [...prevArray, text];
-    });
-  }, [text]);
-
-  useEffect(() => {
-    document.body.style.backgroundColor = "lightblue";
-    return () => {
-      document.body.style.backgroundColor = "";
-    };
-  }, []);
-  console.log(text);
+  const { count, isShow, handleClick, handleDisplay } = useCounter();
+  const {text, inputAlert, array, handleChange, handleAdd } = useInputArray();
+  useBgLightBlue();
 
   return (
     <>
       {isShow ? <h1>{count}</h1> : null}
-      <button 
-        onClick={(e) => handleClick(e)}
-        // onClick={(e) => {handleClick(e), foo}}
-      >
+      <button onClick={(e) => handleClick(e)}>
         ボタン
       </button>
       <button onClick={handleDisplay}>
         {isShow ? "非表示" : "表示"}
       </button>
+
       <input 
         type="text"
         value={text}
@@ -74,6 +35,7 @@ export default function Home() {
           )
         })}
       </ul>
+
       <Main page="index" />
     </>
   )
